backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
python-dotenv==1.0.1
google-genai>=1.47.0
PyMuPDF==1.24.3
pdfplumber==0.11.0
pydantic>=2.7.1
